move module into record linkage directory
