+++ src/pudl/analysis/eia_ferc1_train.py
@@ -1,6 +1,6 @@
 """Create spreadsheets for manually mapping FERC-EIA records and validate matches.
 
-:mod:`pudl.analysis.ferc1_to_eia` uses machine learning to link records from FERC Form 1
+:mod:`pudl.analysis.record_linkage.eia_ferc1_record_linkage` uses machine learning to link records from FERC Form 1
 with records from EIA. While this process is way more efficient and logical
 than a human, it requires a set of hand-compiled training data in order to do it's job.
