+++ src/pudl/analysis/record_linkage/__init__.py
@@ -1,2 +1,2 @@
-"""This module impolements models for various forms of record linkage."""
-from . import classify_plants_ferc1
+"""This module implements models for various forms of record linkage."""
+from . import classify_plants_ferc1, eia_ferc1_record_linkage
